refactor: refactor main navigation
